to finish reserve and devol book

+++ app/Services/Book/BookService.php
@@ -2,16 +2,17 @@
 
 namespace App\Services\Book;
 
+use App\Book\BookLoansStatus;
 use App\Models\Book\Book;
 use App\Repositories\Interfaces\Book\BookInterfaceRepository;
 use App\Repositories\Interfaces\Book\BookLoansInterfaceRepository;
+use Exception;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
 
 class BookService
 {
     public function __construct(
-        protected BookInterfaceRepository $bookRepository,
-        protected BookLoansInterfaceRepository $bookLoansRepository,
+        protected BookInterfaceRepository $bookRepository
     ){}
 
     public function getAll()
@@ -31,7 +32,7 @@ class BookService
         return $this->bookRepository->update($book, $data);
     }
 
-    public function findById(int $id): ?Book
+    public function findById(int $id, ?bool $forLock = false): Book
     {
         $book = $this->bookRepository->findById($id);
 
@@ -42,6 +43,16 @@ class BookService
 
     public function delete(int $id): void
     {
+        $book = $this->findById($id);
+
+        $hasOpenLoans = $book->bookLoans()
+            ->whereIn('status', [
+                BookLoansStatus::ACTIVE->value,
+                BookLoansStatus::LATE->value,
+            ])
+            ->exists();
+
+        if ( $hasOpenLoans ) throw new Exception('Livro com empréstimo em aberto não pode ser excluído.');
 
         $this->bookRepository->delete($id);
     }
